+++ data-r2dbc/src/main/java/io/micronaut/data/r2dbc/operations/DefaultR2dbcRepositoryOperations.java
@@ -40,6 +40,7 @@ import io.micronaut.data.model.PersistentProperty;
 import io.micronaut.data.model.query.JoinPath;
 import io.micronaut.data.model.query.builder.sql.Dialect;
 import io.micronaut.data.model.query.builder.sql.SqlQueryBuilder;
+import io.micronaut.data.model.runtime.AttributeConverterRegistry;
 import io.micronaut.data.model.runtime.DeleteBatchOperation;
 import io.micronaut.data.model.runtime.DeleteOperation;
 import io.micronaut.data.model.runtime.EntityOperation;
@@ -51,7 +52,6 @@ import io.micronaut.data.model.runtime.RuntimeAssociation;
 import io.micronaut.data.model.runtime.RuntimeEntityRegistry;
 import io.micronaut.data.model.runtime.RuntimePersistentEntity;
 import io.micronaut.data.model.runtime.RuntimePersistentProperty;
-import io.micronaut.data.model.runtime.AttributeConverterRegistry;
 import io.micronaut.data.model.runtime.UpdateBatchOperation;
 import io.micronaut.data.model.runtime.UpdateOperation;
 import io.micronaut.data.operations.async.AsyncRepositoryOperations;
@@ -70,6 +70,10 @@ import io.micronaut.data.runtime.mapper.sql.SqlDTOMapper;
 import io.micronaut.data.runtime.mapper.sql.SqlResultEntityTypeMapper;
 import io.micronaut.data.runtime.operations.AsyncFromReactiveAsyncRepositoryOperation;
 import io.micronaut.data.runtime.operations.internal.AbstractSqlRepositoryOperations;
+import io.micronaut.data.runtime.operations.internal.OpContext;
+import io.micronaut.data.runtime.operations.internal.SqlOperation;
+import io.micronaut.data.runtime.operations.internal.StoredAnnotationMetadataSqlOperation;
+import io.micronaut.data.runtime.operations.internal.StoredSqlOperation;
 import io.micronaut.data.runtime.support.AbstractConversionContext;
 import io.micronaut.http.codec.MediaTypeCodec;
 import io.micronaut.transaction.TransactionDefinition;
@@ -379,7 +383,7 @@ final class DefaultR2dbcRepositoryOperations extends AbstractSqlRepositoryOperat
     }
 
     @Override
-    protected int shiftIndex(int i) {
+    public int shiftIndex(int i) {
         return i;
     }
 
@@ -1112,12 +1116,12 @@ final class DefaultR2dbcRepositoryOperations extends AbstractSqlRepositoryOperat
         }
 
         @Override
-        protected void setParameters(Connection connection, Statement stmt, SqlOperation sqlOperation) {
+        protected void setParameters(OpContext<Connection, Statement> context, Connection connection, Statement stmt, SqlOperation sqlOperation) {
             data = data.map(d -> {
                 if (d.vetoed) {
                     return d;
                 }
-                sqlOperation.setParameters(connection, stmt, persistentEntity, d.entity, d.previousValues);
+                sqlOperation.setParameters(context, connection, stmt, persistentEntity, d.entity, d.previousValues);
                 return d;
             });
         }
@@ -1322,12 +1326,12 @@ final class DefaultR2dbcRepositoryOperations extends AbstractSqlRepositoryOperat
         }
 
         @Override
-        protected void setParameters(Connection connection, Statement stmt, SqlOperation sqlOperation) {
+        protected void setParameters(OpContext<Connection, Statement> context, Connection connection, Statement stmt, SqlOperation sqlOperation) {
             entities = entities.map(d -> {
                 if (d.vetoed) {
                     return d;
                 }
-                sqlOperation.setParameters(connection, stmt, persistentEntity, d.entity, d.previousValues);
+                sqlOperation.setParameters(context, connection, stmt, persistentEntity, d.entity, d.previousValues);
                 stmt.add();
                 return d;
             });
